Fix Quantityモジュールをlib下へ切り分け
