+++ lib/quantity.rb
@@ -1,6 +1,4 @@
 module Quantity
-  extend ActiveSupport::Concern
-
   UNITS = [:gram, :gramme, :grams, :grammes, :g]
   DEFAULT_UNIT = :g
   PREFIXES = Alchemist.library.unit_prefixes.each_with_object({}) do |(prefix, value), hash|
@@ -42,28 +40,6 @@ module Quantity
     def unit_exists?(quantity_unit)
       num_and_unit = 1.try(quantity_unit.to_s)
       num_and_unit.class == Alchemist::Measurement && UNITS.include?(num_and_unit.unit_name)
-    end
-  end
-
-  def decimal_quantity
-    Quantity.decimal_quantity(quantity, quantity_unit)
-  end
-
-  def decimal_quantity_was
-    Quantity.decimal_quantity(quantity_was, quantity_unit_was)
-  end
-
-  def string_quantity
-    Quantity.string_quantity(quantity, quantity_unit)
-  end
-
-  def string_quantity_was
-    Quantity.string_quantity(quantity_was, quantity_unit_was)
-  end
-
-  def quantity_unit_exists
-    unless quantity_unit.blank? || Quantity.unit_exists?(quantity_unit)
-      errors.add(:quantity_unit, "\"#{quantity_unit}\" does not exist")
     end
   end
 end
